Cleaning dagostini.C, adding svd for studying response matrix
