+++ dagostini.C
@@ -19,6 +19,7 @@
 #include "TLine.h"
 #include "TStyle.h"
 #include "TLegend.h"
+#include "TDecompSVD.h"
 
 #include "RooUnfold/src/RooUnfold.h"
 #include "RooUnfold/src/RooUnfoldBayes.h"
@@ -37,23 +38,18 @@
 using namespace std;
 
 // Resolution function
-int _jk = 0; // global variable
-bool _jet = false; // global variable
+int _jk = 0;
+bool _jet = false;
 
-Double_t fPtRes(Double_t *x, Double_t *p) {
-
-  return ptresolution(x[0], p[0]);
-}
+Double_t fPtRes(Double_t *x, Double_t *p) { return ptresolution(x[0], p[0]);}
 
 // Ansatz Kernel 
 int cnt_a = 0;
-//const int nk = 4; // number of kernel parameters (excluding pt, eta)
 const int nk = 3; // number of kernel parameters (excluding pt, eta)
+
 Double_t smearedAnsatzKernel(Double_t *x, Double_t *p) {
 
-  if (++cnt_a%1000000==0) {
-    cout << "+" << flush;
-  }
+  if (++cnt_a%1000000==0) cout << "+" << flush;
 
   const double pt = x[0]; // true pT
   const double ptmeas = p[0]; // measured pT
@@ -61,25 +57,20 @@ Double_t smearedAnsatzKernel(Double_t *x, Double_t *p) {
 
   double res = ptresolution(pt, eta+1e-3) * pt;
   const double s = TMath::Gaus(ptmeas, pt, res, kTRUE);
-  //   const double f = p[2] * exp(p[3]/pt) * pow(pt, p[4])
-  //    * pow(1 - pt*cosh(eta) / jp::emax, p[5]);
-  const double f = p[2] * pow(pt, p[3])
-    * pow(1 - pt*cosh(eta) / jp::emax, p[4]);
+  const double f = p[2] * pow(pt, p[3]) * pow(1 - pt*cosh(eta) / jp::emax, p[4]);
   
   return (f * s);
 }
 
-// Smeared Ansatzz
+// Smeared Ansatz
 double _epsilon = 1e-12;
 TF1 *_kernel = 0; // global variable, not pretty but works
 Double_t smearedAnsatz(Double_t *x, Double_t *p) {
 
   const double pt = x[0];
   const double eta = p[0];
-  //const double eta = 0.0;
 
-  if (!_kernel) _kernel = new TF1("_kernel", smearedAnsatzKernel,
-				  1., jp::emax/cosh(eta), nk+2);
+  if (!_kernel) _kernel = new TF1("_kernel", smearedAnsatzKernel, 1., jp::emax/cosh(eta), nk+2);
 
   double res = ptresolution(pt, eta+1e-3) * pt;
   const double sigma = max(0.10, min(res/pt, 0.30));
@@ -89,7 +80,6 @@ Double_t smearedAnsatz(Double_t *x, Double_t *p) {
   //  cout << Form("1pt %10.5f sigma %10.5f ptmin %10.5f ptmax %10.5f eta %10.5f",pt, sigma, ptmin, ptmax, eta) << endl << flush;
   ptmax = min(jp::emax/cosh(eta), ptmax); // safety check
   //  cout << Form("2pt %10.5f sigma %10.5f ptmin %10.5f ptmax %10.5f eta %10.5f",pt, sigma, ptmin, ptmax, eta) << endl << flush;
-
 
   const double par[nk+2] = {pt, eta, p[1], p[2], p[3]};
   _kernel->SetParameters(&par[0]);
@@ -115,15 +105,10 @@ void dagostiniUnfold(string type) {
   // TFile *fin2 = new TFile(Form("output-%s-2c.root",type.c_str()),"READ");
   // TFile *fin2 = new TFile(Form("output-%s-2b.root",type.c_str()),"READ");
   // TFile *fin2 = new TFile(Form("output-%s-2c.root","MC"),"READ");
-  TFile *fin2 = new TFile(jp::dagfile1 ? "output-MC-1.root" : "output-MC-2b.root","READ");
+  TFile *fin2 = new TFile(jp::dagfile1 ? "output-MC-1.root" : "output-MC-2b.root","READ"); assert(fin2 && !fin2->IsZombie());
 
-  assert(fin2 && !fin2->IsZombie());
+  TFile *fout = new TFile(Form("output-%s-3.root",type.c_str()),"RECREATE"); assert(fout && !fout->IsZombie());
 
-  TFile *fout = new TFile(Form("output-%s-3.root",type.c_str()),"RECREATE");
-  assert(fout && !fout->IsZombie());
-
-  _ak7 = (string(jp::algo)=="AK7");
-  if (_ak7) cout << "Using AK7 JER" << endl << flush;
   bool ismc = jp::ismc;
 
   recurseFile(fin, fin2, fout, ismc);
@@ -174,27 +159,10 @@ void recurseFile(TDirectory *indir, TDirectory *indir2, TDirectory *outdir,
     } // inherits from TDirectory
 
     // Found hpt plot: call unfolding routine
-    if (obj->InheritsFrom("TH1") &&
-        (string(obj->GetName())=="hpt" ||
-	 string(obj->GetName())=="hpt_jet" ||
-	 string(obj->GetName())=="hpt_jk1" ||
-	 string(obj->GetName())=="hpt_jk2" ||
-	 string(obj->GetName())=="hpt_jk3" ||
-	 string(obj->GetName())=="hpt_jk4" ||
-	 string(obj->GetName())=="hpt_jk5" ||
-	 string(obj->GetName())=="hpt_jk6" ||
-	 string(obj->GetName())=="hpt_jk7" ||
-	 string(obj->GetName())=="hpt_jk8" ||
-	 string(obj->GetName())=="hpt_jk9" ||
-	 string(obj->GetName())=="hpt_jk10"
-	 )) {
-
+    if (obj->InheritsFrom("TH1") && (string(obj->GetName())=="hpt")) {
       cout << "+" << flush;
 
       _jk = 0;
-      if (TString(obj->GetName()).Contains("hpt_jk")) {
-	assert( sscanf(obj->GetName(), "hpt_jk%d", &_jk) == 1);
-      }
       _jet = TString(obj->GetName()).Contains("hpt_jet");
 
       TH1D *hpt = (TH1D*)obj;
@@ -207,20 +175,6 @@ void recurseFile(TDirectory *indir, TDirectory *indir2, TDirectory *outdir,
 
     } // hpt
 
-    // Try to process friends similarly
-    /*
-    if (obj->InheritsFrom("TH1") &&
-        (string(obj->GetName())=="hpt_ak5calo")) {
-
-      cout << "-" << flush;
-
-      _jk = 0; _jet = false;
-      TH1D *hpt = (TH1D*)obj;
-      TH1D *hpt2 = (TH1D*)indir2->Get("hnlo"); assert(hpt2);
-      if (hpt2)
-        dagostiniUnfold_histo(hpt, hpt2, outdir, ismc, false, "_ak5calo");
-    } // hpt
-    */
   } // while key
 
   curdir->cd();
@@ -282,7 +236,7 @@ void dagostiniUnfold_histo(TH1D *hpt, TH1D *hnlo, TDirectory *outdir,
 
   // Second fit to properly centered graph
   //gnlo2->Fit(fnlo,"QRN");
-  fnlo->SetRange(max(60.,jp::unfptminnlo), min(jp::xmax, jp::emax/cosh(y1)));
+  fnlo->SetRange(max(60.,jp::unfptminnlo), min(jp::xmax, jp::emax/cosh(y1)));   // Fit ranges here...
   cout << "fit to gnlo2" << endl;
   gnlo2->Fit(fnlo,"RN");
   fnlo->SetRange(jp::unfptminnlo, min(jp::xmax, jp::emax/cosh(y1)));
@@ -292,15 +246,15 @@ void dagostiniUnfold_histo(TH1D *hpt, TH1D *hnlo, TDirectory *outdir,
   gpt->SetName(Form("gpt%s",c));
     for (int i = 1; i != hpt->GetNbinsX()+1; ++i) {
  
-    double ptmin = hpt->GetBinLowEdge(i);
-    double ptmax = hpt->GetBinLowEdge(i+1);
-    double y = fnlo->Integral(ptmin, ptmax) / (ptmax - ptmin);
-    double x = fnlo->GetX(y, ptmin, ptmax);
-    double ym = hpt->GetBinContent(i);
-    double ym_err = hpt->GetBinError(i);
-    if (ym>0) {
-      tools::SetPoint(gpt, gpt->GetN(), x, ym, 0., ym_err);
-    }
+      double ptmin = hpt->GetBinLowEdge(i);
+      double ptmax = hpt->GetBinLowEdge(i+1);
+      double y = fnlo->Integral(ptmin, ptmax) / (ptmax - ptmin);
+      double x = fnlo->GetX(y, ptmin, ptmax);
+      double ym = hpt->GetBinContent(i);
+      double ym_err = hpt->GetBinError(i);
+      if (ym>0) {
+	tools::SetPoint(gpt, gpt->GetN(), x, ym, 0., ym_err);
+      }
   } // for i
 
   // Create smeared theory curve
@@ -311,8 +265,7 @@ void dagostiniUnfold_histo(TH1D *hpt, TH1D *hnlo, TDirectory *outdir,
                        fnlo->GetParameter(2), 0, 0);
   cout << "par0 "<< fnlos->GetParameter(0) << "y1 "<< y1<<endl<<flush; 
 
- if (jp::debug)
-    cout << "Calculate forward smearing and unfold hpt" << endl << flush;
+  cout << "Calculate forward smearing and unfold hpt" << endl << flush;
 
   TGraphErrors *gfold_fwd = new TGraphErrors(0);
   gfold_fwd->SetName(Form("gfold_fwd%s",c));
@@ -320,11 +273,12 @@ void dagostiniUnfold_histo(TH1D *hpt, TH1D *hnlo, TDirectory *outdir,
   gcorrpt_fwd->SetName(Form("gcorrpt_fwd%s",c));
   TH1D *hcorrpt_fwd = (TH1D*)hpt->Clone(Form("hcorrpt_fwd%s",c));
 
-  for (int i = 0; i != gpt->GetN(); ++i) {
-    //  for (int i = 1; i != gpt->GetN()+1; ++i) {
+  //  for (int i = 0; i != gpt->GetN(); ++i) {
+  for (int i = 1; i != gpt->GetN()+1; ++i) {
     double x, y, ex, ey;
     tools::GetPoint(gpt, i, x, y, ex, ey);
     double k = fnlo->Eval(x) / fnlos->Eval(x);
+
     if (!TMath::IsNaN(k)) {
 
       tools::SetPoint(gfold_fwd, gfold_fwd->GetN(), x, k, ex, 0.);
@@ -336,10 +290,9 @@ void dagostiniUnfold_histo(TH1D *hpt, TH1D *hnlo, TDirectory *outdir,
     }
   }
 
-
   // Calculate smearing matrix
-  if (jp::debug)
-    cout << "Generating smearing matrix T..." << flush;
+  cout << "Generating smearing matrix T..." << endl << flush;
+
   double tmp_eps = _epsilon;
   _epsilon = 1e-6; // speed up calculations with acceptable loss of precision
 
@@ -347,32 +300,34 @@ void dagostiniUnfold_histo(TH1D *hpt, TH1D *hnlo, TDirectory *outdir,
   outdir->cd();
 
   // Deduce range and binning for true and measured spectra
-  vector<double> vx; // true
-  vector<double> vy; // measured
+  vector<double> vx; // true, gen
+  vector<double> vy; // measured, reco
   for (int i = 1; i != hpt->GetNbinsX()+1; ++i) {
 
     double x = hpt->GetBinCenter(i);
     double x1 = hpt->GetBinLowEdge(i);
     double x2 = hpt->GetBinLowEdge(i+1);
-    double y = hpt->GetBinContent(i);
+    double y = hpt->GetBinContent(i); 
 
-    if (x>=jp::unfptminreco && y>0) {
+    // if (x>=jp::unfptminreco && y>0) {     // CORRECT RANGES?
+    if (x>=jp::unfptmingen && y>0) {
+      
       if (vx.size()==0) vx.push_back(x1);
       vx.push_back(x2);
     }
 
     //   if (x>=jp::fitptmin && y>0) {
-      if (x>=jp::unfptmingen && y>0) {
+     if (x>=jp::unfptminreco && y>0) {
+     //  if (x>=jp::unfptmingen && y>0) {
       if (vy.size()==0) vy.push_back(x1);
       vy.push_back(x2);
     }
   } // for i
 
   // copy over relevant part of hpt
-  TH1D *hreco = new TH1D(Form("hreco%s",c),";p_{T,reco} (GeV)",   // Indices? do we want bin 0 too?
+  TH1D *hreco = new TH1D(Form("hreco%s",c),";p_{T,reco} (GeV)",  
 			 vy.size()-1,&vy[0]);
-  //   for (int i = 1; i != hreco->GetNbinsX()+1; ++i) {  // orig
-   for (int i = 0; i != hreco->GetNbinsX(); ++i) {
+  for (int i = 1; i != hreco->GetNbinsX()+1; ++i) {  // orig
     int j = hpt->FindBin(hreco->GetBinCenter(i));
     double dpt = hpt->GetBinWidth(j);
     hreco->SetBinContent(i, hpt->GetBinContent(j)*dpt);
@@ -382,28 +337,30 @@ void dagostiniUnfold_histo(TH1D *hpt, TH1D *hnlo, TDirectory *outdir,
   // copy over relevant part of hnlo
   TH1D *htrue = new TH1D(Form("htrue%s",c),";p_{T,gen} (GeV)",
 			 vx.size()-1,&vx[0]);
-  //  for (int i = 1; i != htrue->GetNbinsX()+1; ++i) { // orig
-   for (int i = 0; i != htrue->GetNbinsX(); ++i) {
+ for (int i = 1; i != htrue->GetNbinsX()+1; ++i) { 
     int j = hnlo->FindBin(htrue->GetBinCenter(i));
     double dpt = hnlo->GetBinWidth(j);
     htrue->SetBinContent(i, hnlo->GetBinContent(j)*dpt);
     htrue->SetBinError(i, hnlo->GetBinError(j)*dpt);
   }
 
-  TH2D *mt = new TH2D(Form("mt%s",c),"mt;p_{T,reco};p_{T,gen}",
-		      vy.size()-1, &vy[0], vx.size()-1, &vx[0]);
-  TH1D *mx = new TH1D(Form("mx%s",c),"mx;p_{T,gen};#sigma/dp_{T}",
-		      vx.size()-1, &vx[0]);
+   TH2D *mt = new TH2D(Form("mt%s",c),"mt;p_{T,reco};p_{T,gen}",           // Construction: x, y
+		      vy.size()-1, &vy[0],vx.size()-1, &vx[0]
+		       );
+  TH1D *mx = new TH1D(Form("mx%s",c),"mx;p_{T,gen};#sigma/dp_{T}",   
+		      vx.size()-1, &vx[0]);   // "TRUE"
   TH1D *my = new TH1D(Form("my%s",c),"my;p_{T,reco};#sigma/dp_{T}",
-		      vy.size()-1, &vy[0]);
+		      vy.size()-1, &vy[0]);   // RECO
 
+  double mtbinsX = mt->GetNbinsX();  double mtbinsY = mt->GetNbinsY();
+  
   // From http://hepunx.rl.ac.uk/~adye/software/unfold/RooUnfold.html
   // For 1-dimensional true and measured distribution bins Tj and Mi,
   // the response matrix element Rij gives the fraction of events
   // from bin Tj that end up measured in bin Mi.
 
   for (int i = 1; i != mt->GetNbinsX()+1; ++i) {
-
+   
     double ptreco1 = mt->GetXaxis()->GetBinLowEdge(i);
     double ptreco2 = mt->GetXaxis()->GetBinLowEdge(i+1);
     double yreco = fnlo->Integral(ptreco1, ptreco2) / (ptreco2 - ptreco1);
@@ -414,50 +371,77 @@ void dagostiniUnfold_histo(TH1D *hpt, TH1D *hnlo, TDirectory *outdir,
       double ptgen1 = min(jp::emax/cosh(y1), mt->GetYaxis()->GetBinLowEdge(j));
       double ptgen2 = min(jp::emax/cosh(y1), mt->GetYaxis()->GetBinLowEdge(j+1));
 
-      if (ptgen1>=jp::unfptmingen && ptreco>jp::unfptminreco && ptgen1*cosh(y1)<jp::emax) {
+      if (ptgen1>=jp::unfptmingen && ptreco>jp::unfptminreco && ptgen1*cosh(y1)<jp::emax) {  // This results in rows and columns of 0 in mt
 
         fnlos->SetParameter(4, ptgen1);
         fnlos->SetParameter(5, ptgen2);
         // 2D integration over pTreco, pTgen simplified to 1D over pTgen
         mt->SetBinContent(i, j, fnlos->Eval(ptreco) * (ptreco2 - ptreco1));
-        fnlos->SetParameter(4, 0);
+
+	fnlos->SetParameter(4, 0);
         fnlos->SetParameter(5, 0);
       }
     } // for j
   } // for i
 
   for (int j = 1; j != mt->GetNbinsY()+1; ++j) {
-
+  
     double ptgen1 = min(jp::emax/cosh(y1), mt->GetYaxis()->GetBinLowEdge(j));
     double ptgen2 = min(jp::emax/cosh(y1), mt->GetYaxis()->GetBinLowEdge(j+1));
     double ygen = fnlo->Integral(ptgen1, ptgen2);
     mx->SetBinContent(j, ygen);
   }
 
-  for (int i = 1; i != mt->GetNbinsX()+1; ++i) {
+ for (int i = 1; i != mt->GetNbinsX()+1; ++i) {
 
-    double yreco(0);
+    double yrecoprojection(0);
     for (int j = 1; j != mt->GetNbinsY()+1; ++j) {
-      yreco += mt->GetBinContent(i, j);
+    
+      yrecoprojection += mt->GetBinContent(i, j);
     }
-    my->SetBinContent(i, yreco);
+    my->SetBinContent(i, yrecoprojection);
   } // for i
 
   TH2D *mtu = (TH2D*)mt->Clone(Form("mtu%s",c));
   for (int i = 1; i != mt->GetNbinsX()+1; ++i) {
     for (int j = 1; j != mt->GetNbinsY()+1; ++j) {
-      if (mx->GetBinContent(j)!=0) {
+        if (mx->GetBinContent(j)!=0) {
 	mtu->SetBinContent(i, j, mt->GetBinContent(i,j) / mx->GetBinContent(j));
       }
     } // for j
-  } // for i
+  } // for i 
 
+  // Check condition number (statcomm recommendation) of the response matrix mt (TDecompSVD)
+  // cond(K) = sigma_max/max(0,sigma_min), sigmas are singular values of matrix K
 
+  // TH2D to TMatrixD
+  Int_t nbinstotal = mtbinsY*mtbinsX;
+  TMatrixD *K = new TMatrixD(mtbinsY,mtbinsX); // K(rows, cols)
+  TArrayD mtEntries(nbinstotal);
+
+  for (Int_t j = 1; j <= mt->GetNbinsY(); j++) {
+    for (Int_t i = 1; i <= mt->GetNbinsX(); i++) {
+      mtEntries[(j-1)*mtbinsX+i-1] = mtu->GetBinContent(i,j);
+    }
+  }
+
+  K->SetMatrixArray(mtEntries.GetArray());
+  K->Print();
+
+  TDecompSVD *svd = new TDecompSVD(*K);
+  svd->Decompose(); 
+
+  // Get singular values
+  TVectorD singulars = svd->GetSig();
+  singulars.Print();
+  // zeros -> get rid of by eliminaing rows and colums with only 0s?
+  
+  
   // For BinByBin and SVD, need square matrix
   TH2D *mts(0);
   TH1D *mxs(0);
 
-  //  _jet = 0; // added
+   _jet = 0; // added
   
   if (!_jk && !_jet) {
 
@@ -469,8 +453,6 @@ void dagostiniUnfold_histo(TH1D *hpt, TH1D *hnlo, TDirectory *outdir,
     for (int i = 1; i != mts->GetNbinsX()+1; ++i) {
       for (int j = 1; j != mts->GetNbinsY()+1; ++j) {
 
-	//	double x = mts->GetBinCenter(i);   // These two lines give an error: Error in <TH2D::GetBinCenter>: Invalid method for a 2-d histogram - return a NaN
-	// double y = mts->GetBinCenter(j);
 	double x = ((TAxis*)mts->GetXaxis())->GetBinCenter(i);
 	double y = ((TAxis*)mts->GetYaxis())->GetBinCenter(j);
 	int i2 = mt->GetXaxis()->FindBin(x);
@@ -494,7 +476,7 @@ void dagostiniUnfold_histo(TH1D *hpt, TH1D *hnlo, TDirectory *outdir,
   if (jp::debug)
     cout << "done." << endl << flush;
 
-  /*
+ 
   outdir->cd();
   if (!_jk) {
     hreco->Write();
@@ -503,7 +485,7 @@ void dagostiniUnfold_histo(TH1D *hpt, TH1D *hnlo, TDirectory *outdir,
     mt->Write();
     mtu->Write();
   }
-  */
+ 
 
   // Now to actual unfolding business with the d'Agostini method
   if (jp::debug)
@@ -529,7 +511,7 @@ void dagostiniUnfold_histo(TH1D *hpt, TH1D *hnlo, TDirectory *outdir,
   //                 const char* name= 0, const char* title= 0);
   RooUnfoldBayes *uBayes = new RooUnfoldBayes(uResp, hreco, 4);
 
-  if (jp::debug)
+// if (jp::debug)
     uBayes->Print();
 
   TH1D *hTrueBayes = (TH1D*)uBayes->Hreco(RooUnfold::kCovariance);
@@ -584,34 +566,32 @@ void dagostiniUnfold_histo(TH1D *hpt, TH1D *hnlo, TDirectory *outdir,
 
     cout << "Moving to SVD" << endl; 
     
-    bool _svd = true;
-    if (_svd) {
-      int kreg = int(vy.size()/2);
-      int ntoys = 0;// default 1000
-      RooUnfoldSvd *uSVD = new RooUnfoldSvd(uResps, hreco, kreg, ntoys);
-      TH1D *hTrueSVD = (TH1D*)uSVD->Hreco();//RooUnfold::kCovariance);
-      assert(hTrueSVD);
-      hcorrpt_svd = (TH1D*)hTrueSVD->Clone(Form("hcorrpt_svd%s",c));
-      delete uSVD; // ensure static members get destroyed before next instance
+    int kreg = int(vy.size()/2);
+    int ntoys = 0;// default 1000
+    RooUnfoldSvd *uSVD = new RooUnfoldSvd(uResps, hreco, kreg, ntoys);
+    TH1D *hTrueSVD = (TH1D*)uSVD->Hreco();//RooUnfold::kCovariance);
+    assert(hTrueSVD);
+    hcorrpt_svd = (TH1D*)hTrueSVD->Clone(Form("hcorrpt_svd%s",c));
+    delete uSVD; // ensure static members get destroyed before next instance
 
-      hcorrpt_svd = (TH1D*)hpt->Clone(Form("hcorrpt_svd%s",c));
-      hcorrpt_svd->Reset();
-      for (int i = 1; i != hTrueSVD->GetNbinsX()+1; ++i) {
+    hcorrpt_svd = (TH1D*)hpt->Clone(Form("hcorrpt_svd%s",c));
+    hcorrpt_svd->Reset();
+    for (int i = 1; i != hTrueSVD->GetNbinsX()+1; ++i) {
 
-	int j = hpt->FindBin(hTrueSVD->GetBinCenter(i));
-	hcorrpt_svd->SetBinContent(j, hTrueSVD->GetBinContent(i));
-	hcorrpt_svd->SetBinError(j, hTrueSVD->GetBinError(i));
+      int j = hpt->FindBin(hTrueSVD->GetBinCenter(i));
+      hcorrpt_svd->SetBinContent(j, hTrueSVD->GetBinContent(i));
+      hcorrpt_svd->SetBinError(j, hTrueSVD->GetBinError(i));
       }
-    }
-    else  {
-      hcorrpt_svd = (TH1D*)hcorrpt_bin->Clone(Form("hcorrpt_svd%s",c));
-      hcorrpt_svd->Reset();
-    }
-  } // !_jk
+   } // !_jk
+
+  // TUnfold.
+
+  // cout << "Moving to TUnfold" << endl;
+
+
 
   if (jp::debug)
-    cout << "done." << endl << flush;
-
+   cout << "done." << endl << flush;
 
   // Store "unfolding correction" (unfolded / original) and corrected graph
   //cout << "Store graphs..." << endl << flush;
@@ -759,7 +739,9 @@ void dagostiniUnfold_histo(TH1D *hpt, TH1D *hnlo, TDirectory *outdir,
     gfold_dag->Write("gfold");
 
     // Fit functions
-    //uResp->Write();
+    uResp->Write();
+    fnlo->SetRange(jp::unfptminnlo, min(jp::xmax, jp::emax/cosh(y1)));
+    fnlo->SetNpx(1000); // otherwise ugly on log x-axis after write
     fnlo->Write();
     // Calculating points for fs is taking significant time,
     // and even got stuck at some point when too far out of the range
@@ -777,7 +759,8 @@ void dagostiniUnfold_histo(TH1D *hpt, TH1D *hnlo, TDirectory *outdir,
     htrue->Write();
     mt->Write();
     mtu->Write();
-
+    //    K->Write();
+    
     // Alternative methods
     gfold_dag->Write();
     gfold_fwd->Write();
@@ -812,247 +795,7 @@ void dagostiniUnfold_histo(TH1D *hpt, TH1D *hnlo, TDirectory *outdir,
     hcorrpt_dag->Write("hcorrpt");
   }
 
+  
 } // dagostiniUnfold_histo
 
-
-void drawDagostini(string type) {
-
-  TDirectory *curdir = gDirectory;
-  setTDRStyle();
-
-  TFile *f = new TFile(Form("output-%s-3.root",type.c_str()),"READ");
-  assert(f && !f->IsZombie());
-
-  assert(f->cd("Standard"));
-  f->cd("Standard");
-  TDirectory *din = f->GetDirectory("Standard"); assert(din);
-  curdir->cd();
-
-  TCanvas *c1 = new TCanvas("c1","c1",1200,1200);   // 800
-  //c1->SetTopMargin(0.20);
-  c1->Divide(3,3);//,0.1,0.00); 3,2
-
-  TCanvas *c1b = new TCanvas("c1b","c1b",600,600);
-
-  TCanvas *c2 = new TCanvas("c2","c2",1200,1200);
-  c2->SetTopMargin(0.10);
-  c2->Divide(3,3,-1,-1);
-
-  TCanvas *c3 = new TCanvas("c3","c3",1200,1200);
-  c3->SetTopMargin(0.10);
-  c3->Divide(3,3,-1,-1);
-
-  TH1D *h = new TH1D("h",";p_{T} (GeV);Unfolding correction",
-		     int(jp::xmax-jp::xmin),jp::xmin,jp::xmax);
-  h->SetMinimum(0.45);
-  h->SetMaximum(1.15);
-  h->GetXaxis()->SetMoreLogLabels();
-  h->GetXaxis()->SetNoExponent();
-
-  TLegend *leg = new TLegend(0.25,0.74,0.45,0.97,"","brNDC");
-  leg->SetFillStyle(kNone);
-  leg->SetBorderSize(0);
-  leg->SetTextSize(0.045);
-
-  TLatex *tex = new TLatex();
-  tex->SetTextSize(0.045);
-
-  const int ny = 7;
-  double y1 = 0; double y2 = 0;
-  for (int iy = 0; iy != ny; ++iy) {
-    if (iy==6) {y1 = 3.2; y2 = 4.7;}
-    else {y1 = 0.5*iy;  y2 = 0.5*(iy+1);}
-
-    assert(din->cd(Form("Eta_%1.1f-%1.1f",y1,y2)));
-    din->cd(Form("Eta_%1.1f-%1.1f",y1,y2));
-    TDirectory *d = din->GetDirectory(Form("Eta_%1.1f-%1.1f",y1,y2)); assert(d);
-
-    TH1D *hreco = (TH1D*)d->Get("hreco"); assert(hreco);
-    TH2D *h2resp = (TH2D*)d->Get("mtu"); assert(h2resp);
-    TGraphErrors *gfwd = (TGraphErrors*)d->Get("gfold_fwd"); assert(gfwd);
-    TGraphErrors *gbin = (TGraphErrors*)d->Get("gfold_bin"); assert(gbin);
-    TGraphErrors *gsvd = (TGraphErrors*)d->Get("gfold_svd"); assert(gsvd);
-    TGraphErrors *gbayes = (TGraphErrors*)d->Get("gfold"); assert(gbayes);
-    curdir->cd();
-
-    // Find the low end of reconstructed data used in unfolding
-    double ptmin(0);
-    for (int i = 1; i != hreco->GetNbinsX()+1 && !ptmin; ++i) {
-      if (hreco->GetBinContent(i)>0) ptmin = hreco->GetBinLowEdge(i);
-    }
-
-    c1->cd(iy+1);
-    gPad->SetLogx();
-    gPad->SetLogy();
-    gStyle->SetPalette(1);
-    gPad->SetRightMargin(0.10);
-    h2resp->GetXaxis()->SetMoreLogLabels();
-    h2resp->GetXaxis()->SetNoExponent();
-    h2resp->GetXaxis()->SetTitle("Measured p_{T,reco} (GeV)");
-    h2resp->GetXaxis()->SetTitleOffset(1.5);
-    h2resp->GetYaxis()->SetMoreLogLabels();
-    h2resp->GetYaxis()->SetNoExponent();
-    h2resp->GetYaxis()->SetTitle("True p_{T,gen} (GeV)");
-    h2resp->GetYaxis()->SetTitleOffset(2.0);
-    h2resp->GetZaxis()->SetRangeUser(1e-3,0.9999);//1.05);
-    h2resp->DrawClone("COLZ");
-    //h2resp->Draw("SAMEBOX");
-
-    tex->SetTextSize(0.045);
-    tex->SetTextAlign(31); // align right
-    tex->SetNDC(kTRUE);
-    tex->DrawLatex(0.8, 0.2, y1==0 ? "|y|<0.5" : Form("%1.1f<|y|<%1.1f",y1,y2));
-
-    if (y1==0) {
-      c1b->cd();
-      gPad->SetLogx();
-      gPad->SetLogy();
-      gPad->SetRightMargin(0.13);//0.11);
-      gPad->SetBottomMargin(0.14);//0.10);
-      gPad->SetLeftMargin(0.18);//0.10);
-      //h2resp->GetXaxis()->SetTitleOffset(1.4);
-      //TH2D *h2 = (TH2D*)h2resp->DrawClone("COLZ");
-
-      // copy TH2D over to a fresh one to get default graphics style
-      TH2D *h2 = new TH2D("h2",";Measured Jet p_{T} (GeV);"
-			  "True Jet p_{T} (GeV)",
-			  h2resp->GetNbinsX(),
-			  h2resp->GetXaxis()->GetXbins()->GetArray(),
-			  h2resp->GetNbinsY(),
-			  h2resp->GetYaxis()->GetXbins()->GetArray());
-      for (int i = 1; i != h2->GetNbinsX()+1; ++i) {
-	for (int j = 1; j != h2->GetNbinsY()+1; ++j) {
-	  h2->SetBinContent(i, j, h2resp->GetBinContent(i, j));
-	  //h2->SetBinError(i, j, h2resp->GetBinError(i, j));
-	} // for j
-      } // for i
-      h2->Draw("COLZ");
-      h2->GetXaxis()->SetMoreLogLabels();
-      h2->GetXaxis()->SetNoExponent();
-      //h2->GetYaxis()->SetMoreLogLabels();
-      h2->GetYaxis()->SetNoExponent();
-      h2->GetYaxis()->SetTitleOffset(1.5);
-      h2->GetZaxis()->SetRangeUser(1e-3,0.9999);
-
-      //    h2->GetXaxis()->SetRangeUser(jp::fitptmin,jp::xmax);//1327.);
-      // h2->GetYaxis()->SetRangeUser(jp::recopt,jp::xmax);
-      
-      h2->GetXaxis()->SetRangeUser(jp::unfptminreco,jp::xmax);//1327.);
-      h2->GetYaxis()->SetRangeUser(jp::unfptmingen,jp::xmax);
-      //h2resp->Draw("SAME COLZ");
-
-      //tex->DrawLatex(0.8, 0.15, _algo=="AK7" ?
-      tex->DrawLatex(0.8, 0.20, string(jp::algo)=="AK7" ?
-		     "Anti-k_{T} R=0.8, |y|<0.5" :
-		     "Anti-k_{T} R=0.4, |y|<0.5");
-      //cmsPrel(_lumi);
-      //cmsPrel(0); // simulation
-    }
-
-    c2->cd(iy+1);
-    gPad->SetLogx();
-
-    h->SetMinimum(0.1);
-    h->SetMaximum(1.15);
-    h->SetXTitle(iy==ny-1 ? "p_{T} (GeV)" : "");
-    h->SetYTitle(iy==0 ? "Unfolding correction" : "");
-    h->DrawClone("AXIS");
-
-    TLine *l = new TLine();
-    l->SetLineStyle(kDotted);
-    l->DrawLine(ptmin,0.45,ptmin,1.15);
-    l->SetLineStyle(kDashed);
-    //l->DrawLine(56,0.45,56,1.15); // v2
-    l->DrawLine(jp::xminpas,0.45,jp::xminpas,1.15);
-
-    gfwd->SetName("gfwd");
-    gfwd->SetLineWidth(2);
-    gfwd->SetLineColor(kRed);
-    gfwd->Draw("SAMEL");
-
-    gbin->SetName("gbin");
-    gbin->SetMarkerStyle(kOpenSquare);
-    gbin->SetMarkerColor(kGreen+1);
-    gbin->SetLineColor(kGreen+1);
-     gbin->Draw("SAMEPz");
-
-    gbayes->SetName("gbayes");
-    gbayes->SetMarkerStyle(kFullCircle);
-    gbayes->Draw("SAMEP");
-
-    gsvd->SetName("gsvd");
-    gsvd->SetMarkerStyle(kOpenDiamond);
-    gsvd->SetMarkerColor(kCyan+1);
-    gsvd->SetLineColor(kCyan+1);
-     gsvd->Draw("SAMEP");
-
-    tex->SetTextSize(iy<3 ? 0.053 : 0.045);
-    tex->SetTextAlign(21); // align middle
-    tex->SetNDC(kFALSE);
-    tex->DrawLatex(150, 0.5, y1==0 ? "|y|<0.5" : Form("%1.1f<|y|<%1.1f",y1,y2));
-
-    if (iy==2) {
-      leg->AddEntry(gbayes,"RooUnfoldBayes","P");
-      leg->AddEntry(gbin,"RooUnfoldBinByBin","P");
-       leg->AddEntry(gsvd,"RooUnfoldSvd","P");
-      leg->AddEntry(gfwd,"Forward smearing","L");
-      leg->Draw();
-    }
-
-    c3->cd(iy+1);
-    gPad->SetLogx();
-
-    h->SetMinimum(0.90+0.0001);
-    h->SetMaximum(1.10-0.0001);
-    h->SetXTitle(iy==ny-1 ? "p_{T} (GeV)" : "");
-    h->SetYTitle(iy==0 ? "Correction / Forward smearing" : "");
-    h->DrawClone("AXIS");
-
-    l->SetLineStyle(kDotted);
-    l->DrawLine(ptmin,0.9,ptmin,1.1);
-    l->SetLineStyle(kDashed);
-    //l->DrawLine(56,0.9,56,1.1); // v2
-    l->DrawLine(jp::xminpas,0.9,jp::xminpas,1.1);
-
-    TGraphErrors *grfwd = tools::ratioGraphs(gfwd, gfwd);
-    grfwd->Draw("SAMEL");
-    TGraphErrors *grbin = tools::ratioGraphs(gbin, gfwd);
-    grbin->Draw("SAMEPz");
-    TGraphErrors *grbayes = tools::ratioGraphs(gbayes, gfwd);
-    grbayes->Draw("SAMEPz");
-    TGraphErrors *grsvd = tools::ratioGraphs(gsvd, gfwd);
-    grsvd->Draw("SAMEPz");
-
-    tex->DrawLatex(150,0.91, y1==0 ? "|y|<0.5" : Form("%1.1f<|y|<%1.1f",y1,y2));
-
-    if (iy==2) leg->Draw();
-
-  }
-
-  const char *a = string(jp::algo).c_str();
-  const char *t = type.c_str();
-
-  c1->cd(3);
-  tex->SetTextSize(0.045);
-  tex->SetNDC(kTRUE);
-  tex->DrawLatex(0.35, 0.85, Form("%s %s",t,a));
-  c1->cd(0);
-  //cmsPrel(type=="DATA" ? _lumi : 0, true);
-  c1->SaveAs(Form("pdf/roounfold_matrix_%s_%s.pdf",a,t));
-  c1b->SaveAs(Form("pdf/roounfold_matrix0_%s_%s.pdf",a,t));
-
-  c2->cd(2);
-  tex->SetTextSize(0.053);
-  tex->DrawLatex(0.50, 0.85, Form("%s %s",t,a));
-  c2->cd(0);
-  //cmsPrel(type=="DATA" ? jp::lumi : 0, true);
-  c2->SaveAs(Form("pdf/roounfold_comparison_%s_%s.pdf",a,t));
-
-  c3->cd(2);
-  tex->DrawLatex(0.50, 0.85, Form("%s %s",t,a));
-  c3->cd(0);
-  //cmsPrel(type=="DATA" ? jp::lumi : 0, true);
-  c3->SaveAs(Form("pdf/roounfold_ratiotofwd_%s_%s.pdf",a,t));
-
-
-} // drawDagostini
+
